Simple tests for the conversion functions
