+++ src/Celeste.jl
@@ -2,6 +2,7 @@ module Celeste
 
 require(joinpath(Pkg.dir("Celeste"), "src", "Util.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "CelesteTypes.jl"))
+require(joinpath(Pkg.dir("Celeste"), "src", "Constrain.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "ModelInit.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "SDSS.jl"))
 require(joinpath(Pkg.dir("Celeste"), "src", "ElboDeriv.jl"))
@@ -13,6 +14,8 @@ import SDSS
 import ElboDeriv
 import OptimizeElbo
 import ModelInit
+import SampleData
+import Constrain
 
 using CelesteTypes
