+++ src/CelesteTypes.jl
@@ -20,15 +20,12 @@ export zero_sensitive_float, const_sensitive_param, clear!
 export ParamIndex, ids, ids_free, all_params, all_params_free
 export star_pos_params, galaxy_pos_params
 export D, B, Ia
-export unconstrain_vp, constrain_vp!, rect_unconstrain_vp, rect_constrain_vp!
 
 using Util
 
 import FITSIO
 import Distributions
 import WCSLIB
-
-
 
 const band_letters = ['u', 'g', 'r', 'i', 'z']
 
@@ -348,146 +345,6 @@ end
 
 const ids = get_param_ids()
 const ids_free = get_unconstrained_param_ids()
-
-function rect_unconstrain_vp!(vp::VariationalParams, vp_free::RectVariationalParams)
-    # Convert a constrained to an unconstrained variational parameterization
-    # that does not use logit or exp.
-
-    S = length(vp)
-    for s = 1:S
-        # Variables that are unaffected by constraints:
-        vp_free[s][ids_free.mu] = vp[s][ids.mu]
-        vp_free[s][ids_free.theta] = vp[s][ids.theta]
-        vp_free[s][ids_free.rho] = vp[s][ids.rho]
-        vp_free[s][ids_free.phi] = vp[s][ids.phi]
-        vp_free[s][ids_free.sigma] = vp[s][ids.sigma]
-        vp_free[s][ids_free.kappa] = vp[s][ids.kappa]
-        vp_free[s][ids_free.beta] = vp[s][ids.beta]
-        vp_free[s][ids_free.lambda] = vp[s][ids.lambda]
-        vp_free[s][ids_free.gamma_free] = vp[s][ids.gamma]
-        vp_free[s][ids_free.zeta_free] = vp[s][ids.zeta]
-
-        # Simplicial constriants.  The original script used "chi" to only
-        # refer to the probability of being a galaxy, which is now the
-        # second component of chi.
-        vp_free[s][ids_free.chi_free[1]] = vp[s][ids.chi[2]]
-    end
-end
-
-
-function unconstrain_vp!(vp::VariationalParams, vp_free::FreeVariationalParams)
-    # Convert a constrained to an unconstrained variational parameterization
-    # on all of the real line.
-    S = length(vp)
-    for s = 1:S
-        # Variables that are unaffected by constraints:
-        vp_free[s][ids_free.mu] = vp[s][ids.mu]
-        vp_free[s][ids_free.theta] = vp[s][ids.theta]
-        vp_free[s][ids_free.rho] = vp[s][ids.rho]
-        vp_free[s][ids_free.phi] = vp[s][ids.phi]
-        vp_free[s][ids_free.sigma] = vp[s][ids.sigma]
-        vp_free[s][ids_free.kappa] = vp[s][ids.kappa]
-        vp_free[s][ids_free.beta] = vp[s][ids.beta]
-        vp_free[s][ids_free.lambda] = vp[s][ids.lambda]
-
-        # Simplicial constriants.  The original script used "chi" to only
-        # refer to the probability of being a galaxy, which is now the
-        # second component of chi.
-        vp_free[s][ids_free.chi_free[1]] = Util.inv_logit(vp[s][ids.chi[2]])
-
-        # Positivity constraints
-        vp_free[s][ids_free.gamma_free] = log(vp[s][ids.gamma])
-        vp_free[s][ids_free.zeta_free] = log(vp[s][ids.zeta]) 
-    end
-end
-
-function rect_constrain_vp!(vp_free::RectVariationalParams, vp::VariationalParams)
-    # Convert an unconstrained to an constrained variational parameterization
-    # where we don't use exp or logit.
-
-    S = length(vp_free)
-    for s = 1:S
-        # The default is everything being the same.
-
-        # Maybe something like this instead:
-        #
-        # for id_symbol in names(ids)
-        #     if id_symbol != convert(Symbol, chi)
-        #         vp[s][ids.(id_symbol)] = vp_free[s][ids_free.(id_symbol)]
-        # end
-
-        vp[s][ids.mu] = vp_free[s][ids_free.mu]
-        vp[s][ids.theta] = vp_free[s][ids_free.theta]
-        vp[s][ids.rho] = vp_free[s][ids_free.rho]
-        vp[s][ids.phi] = vp_free[s][ids_free.phi]
-        vp[s][ids.sigma] = vp_free[s][ids_free.sigma]
-        vp[s][ids.kappa] = vp_free[s][ids_free.kappa]
-        vp[s][ids.beta] = vp_free[s][ids_free.beta]
-        vp[s][ids.lambda] = vp_free[s][ids_free.lambda]
-        vp[s][ids.gamma] = vp_free[s][ids_free.gamma_free]
-        vp[s][ids.zeta] = vp_free[s][ids_free.zeta_free]
-
-        # Simplicial constriants.
-        vp[s][ids.chi[2]] = vp_free[s][ids_free.chi_free[1]]
-        vp[s][ids.chi[1]] = 1.0 - vp[s][ids.chi[2]]
-    end
-end
-
-function constrain_vp!(vp_free::FreeVariationalParams, vp::VariationalParams)
-    # Convert an unconstrained to an constrained variational parameterization.
-    S = length(vp_free)
-    for s = 1:S
-        # The default is everything being the same.
-        vp[s][ids.mu] = vp_free[s][ids_free.mu]
-        vp[s][ids.theta] = vp_free[s][ids_free.theta]
-        vp[s][ids.rho] = vp_free[s][ids_free.rho]
-        vp[s][ids.phi] = vp_free[s][ids_free.phi]
-        vp[s][ids.sigma] = vp_free[s][ids_free.sigma]
-        vp[s][ids.kappa] = vp_free[s][ids_free.kappa]
-        vp[s][ids.beta] = vp_free[s][ids_free.beta]
-        vp[s][ids.lambda] = vp_free[s][ids_free.lambda]
-
-        # Simplicial constriants.
-        vp[s][ids.chi[2]] = Util.logit(vp_free[s][ids_free.chi_free[1]])
-        vp[s][ids.chi[1]] = 1.0 - vp[s][ids.chi[2]]
-
-         # Positivity constraints
-        vp[s][ids.gamma] = exp(vp_free[s][ids_free.gamma_free])
-        vp[s][ids.zeta] = exp(vp_free[s][ids_free.zeta_free])
-    end
-end
-
-function unconstrain_vp(vp::VariationalParams)
-    # Convert a constrained to an unconstrained variational parameterization.
-    S = length(vp)
-    vp_free = [ zeros(ids_free.size) for s = 1:S]
-    unconstrain_vp!(vp, vp_free)
-    vp_free
-end
-
-function constrain_vp(vp_free::FreeVariationalParams)
-    # Convert an unconstrained to an constrained variational parameterization.
-    S = length(vp_free)
-    vp = [ zeros(ids.size) for s = 1:S]
-    constrain_vp!(vp_free, vp)
-    vp
-end
-
-function rect_unconstrain_vp(vp::VariationalParams)
-    # Convert a constrained to an unconstrained variational parameterization.
-    S = length(vp)
-    vp_free = [ zeros(ids_free.size) for s = 1:S]
-    rect_unconstrain_vp!(vp, vp_free)
-    vp_free
-end
-
-function rect_constrain_vp(vp_free::RectVariationalParams)
-    # Convert an unconstrained to an constrained variational parameterization.
-    S = length(vp_free)
-    vp = [ zeros(ids.size) for s = 1:S]
-    rect_constrain_vp!(vp_free, vp)
-    vp
-end
 
 
 const all_params = [1:ids.size]
